Update Debugger.py and finetune_script2.py

+++ Finetuning/finetune_script2.py
@@ -47,7 +47,7 @@ if __name__ == "__main__":
                           lr = 3e-4, 
                           sigma = 7.0, 
                           target_reward = 1.0, 
-                          train_goal = np.array([[-2.5, -2.5]], dtype = int),
+                          train_goal = np.array([[-2.5, -2.5]], dtype = float),
                           rollout_goal = np.array([[6, 1]], dtype = int))
     
     TrainKernelConfig = Train_Kernel_Config(
